+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 Inited :)
 
 ### 0.1.1
+
+* `Repos`
+    * `Common`
+        * Extension `ReadStandardCRUDRepo#doWithAll` has been added